+++ spring-data-elasticsearch-rest/src/main/java/com/github/ydespreaux/spring/data/elasticsearch/core/ReactiveElasticsearchOperations.java
@@ -111,6 +111,8 @@ public interface ReactiveElasticsearchOperations {
         return this.createRolloverIndex(null, aliasWriter, indexName);
     }
 
+    <T> Boolean createIndexSync(Class<T> clazz);
+
     /**
      *
      * @param aliasReader