+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ### Added
 - Dependency spring-boot version from 2.1.0 to 2.1.3
 - Dependency elasticsearch version from 6.4.2 to 6.5.0
-- Reactive repository (o)
+- Reactive repository
 
 ### Changed